+++ src/mimefactory.rs
@@ -1733,11 +1733,11 @@ mod tests {
         }
         async fn get_subject(
             t: &TestContext,
-            sent: crate::test_utils::SentMessage,
+            sent: crate::test_utils::SentMessage<'_>,
         ) -> Result<String> {
             let parsed_subject = t.parse_msg(&sent).await.get_subject().unwrap();
 
-            let sent_msg = Message::load_from_db(t, sent.sender_msg_id).await?;
+            let sent_msg = sent.load_from_db().await;
             assert_eq!(parsed_subject, sent_msg.subject);
 
             Ok(parsed_subject)